sirma-platform/platform/seip-parent/platform/emf-semantic/emf-semantic-impl/src/main/java/com/sirma/itt/emf/semantic/search/operation/StartsWithSearchOperation.java
package com.sirma.itt.emf.semantic.search.operation;

import com.sirma.itt.seip.domain.search.SearchOperation;
import com.sirma.itt.seip.domain.search.tree.Rule;
import com.sirma.itt.seip.plugin.Extension;

/**
 * Builds a SPARQL query for a starts with statement from a provided {@link Rule}.
 * 
 * @author Mihail Radkov
 */
@Extension(target = SearchOperation.SPARQL_SEARCH_OPERATION, order = 70)
public class StartsWithSearchOperation implements SearchOperation {

	private static final String STARTS_WITH_OPERATION = "starts_with";

	@Override
	public boolean isApplicable(Rule rule) {
		return STARTS_WITH_OPERATION.equalsIgnoreCase(rule.getOperation());
	}

	@Override
	public void buildOperation(StringBuilder builder, Rule rule) {
		SemanticSearchOperationUtils.appendStringTripleAndFilter(builder, rule, true, false);
	}

}
